+++ src/Figure/Figure.types.ts
@@ -1,4 +1,4 @@
-export interface FigureProps {
+export type FigureProps = {
   /**
    * The path/url to the image
    */
@@ -26,4 +26,4 @@ export interface FigureProps {
    * The scale of the image, similar to size, but can't affect the aspect ratio
    */
   scale?: number;
-}
+};
